# improve performance of `unstructured_2d_to_1d_curve` for linear `P4estMesh`es in 2D