This is basically #2334 in 2D for the `P4estMesh`. It depends on #2326 being merged first since I built on that branch.

After compilation, I get with this PR

```julia
julia> using Revise; using Trixi, Plots

julia> begin
       equations = CompressibleEulerEquations2D(1.4)
       solver = DGSEM(polydeg = 3, surface_flux = flux_lax_friedrichs)
       coordinates_min = (-1.0, -1.0)
       coordinates_max = (+1.0, +1.0)
       initial_refinement_level = 6

       mesh_tree = TreeMesh(coordinates_min, coordinates_max;
                            n_cells_max = 10^5, initial_refinement_level)
       trees_per_dimension = (1, 1)
       mesh_t8code = T8codeMesh(trees_per_dimension; polydeg = 1,
                                coordinates_min, coordinates_max, initial_refinement_level)
       mesh_p4est = P4estMesh(trees_per_dimension; polydeg = 1,
                              coordinates_min, coordinates_max, initial_refinement_level)
       cells_per_dimension = (2, 2) .^ initial_refinement_level
       mesh_structured = StructuredMesh(cells_per_dimension, coordinates_min, coordinates_max)

       ic = initial_condition_constant
       ode_tree = let mesh = mesh_tree
           semi = SemidiscretizationHyperbolic(mesh, equations, ic, solver)
           ode = semidiscretize(semi, (0.0, 0.1))
       end

       ode_t8code = let mesh = mesh_t8code
           semi = SemidiscretizationHyperbolic(mesh, equations, ic, solver)
           ode = semidiscretize(semi, (0.0, 0.1))
       end

       ode_p4est = let mesh = mesh_p4est
           semi = SemidiscretizationHyperbolic(mesh, equations, ic, solver)
           ode = semidiscretize(semi, (0.0, 0.1))
       end

       ode_structured = let mesh = mesh_structured
           semi = SemidiscretizationHyperbolic(mesh, equations, ic, solver)
           ode = semidiscretize(semi, (0.0, 0.1))
       end

       x = range(0, 1, length = 1000)
       curve = vcat(x', x')
       pd_tree = @time PlotData1D(ode_tree.u0, ode_tree.p; curve)
       pd_t8code = @time PlotData1D(ode_t8code.u0, ode_t8code.p; curve)
       pd_p4est = @time PlotData1D(ode_p4est.u0, ode_p4est.p; curve)
       pd_structured = @time PlotData1D(ode_structured.u0, ode_structured.p; curve)

       fig = plot(pd_tree, label = "tree")
       plot!(fig, pd_t8code, label = "t8code", linestyle = :dash)
       plot!(fig, pd_p4est, label = "p4est", linestyle = :dot)
       plot!(fig, pd_structured, label = "structured", linestyle = :dashdot)
       plot!(fig, legend = true)
       end
  0.008353 seconds (27 allocations: 4.169 MiB)
  0.305755 seconds (197 allocations: 8.338 MiB)
  0.000386 seconds (16 allocations: 71.922 KiB)
  0.224093 seconds (197 allocations: 8.338 MiB)
```


On `main`, I get the timings

```
  0.009409 seconds (27 allocations: 4.169 MiB)
  7.869152 seconds (258.18 M allocations: 19.241 GiB, 10.45% gc time, 0.31% compilation time)
  7.786630 seconds (258.14 M allocations: 19.238 GiB, 10.34% gc time)
  6.736877 seconds (222.57 M allocations: 16.588 GiB, 9.79% gc time)
```

The differences for the `T8codeMesh` and the `StructuredMesh` come from the improvements made in #2326. The significant improvement for the `P4estMesh` comes from the algorithmic improvement made in this PR.